fix: delete waitXseconds, move shinyVault to animations, remove handle and handleShadow duplicate animation
