+++ src/animations/shinyVault.ts
